hide Roles and Invites for non-admins + cleanup
