+++ components/members/components/MemberProfile/components/ProfileWidgets/components/MemberPropertiesWidget/MemberPropertiesFormDialog.tsx
@@ -9,7 +9,7 @@ import { Button } from 'components/common/Button';
 import Dialog from 'components/common/DatabaseEditor/components/dialog';
 import ScrollableWindow from 'components/common/PageLayout/components/ScrollableWindow';
 import { useRequiredMemberPropertiesForm } from 'components/members/hooks/useRequiredMemberProperties';
-import Legend from 'components/settings/Legend';
+import Legend from 'components/settings/components/Legend';
 import type { FormFieldValue } from 'lib/forms/interfaces';
 import type { UpdateMemberPropertyValuePayload } from 'lib/members/interfaces';
