+++ components/settings/components/Legend.tsx
