+++ components/settings/config.tsx
@@ -9,6 +9,7 @@ import ManageAccountsIcon from '@mui/icons-material/ManageAccounts';
 import NotificationsOutlinedIcon from '@mui/icons-material/NotificationsOutlined';
 import SettingsIcon from '@mui/icons-material/WorkOutline';
 import MedalIcon from '@mui/icons-material/WorkspacePremium';
+import type { ReactNode } from 'react';
 import { FaBriefcase } from 'react-icons/fa';
 
 import { CharmsLogo } from 'components/common/CharmsLogo';
@@ -16,10 +17,24 @@ import { ProposalIcon } from 'components/common/PageIcon';
 
 export const premiumSettingSections = ['roles', 'api'] as const;
 
-export const SPACE_SETTINGS_TABS = [
+type SpaceSettingsSection =
+  | 'space'
+  | 'roles'
+  | 'invites'
+  | 'import'
+  | 'api'
+  | 'subscription'
+  | 'credentials'
+  | 'proposals'
+  | 'integrations'
+  | 'notifications';
+
+export type SpaceSettingsTab = { icon: ReactNode; path: SpaceSettingsSection; label: string; adminOnly?: boolean };
+
+export const SPACE_SETTINGS_TABS: SpaceSettingsTab[] = [
   { icon: <SettingsIcon fontSize='small' />, path: 'space', label: 'Overview' },
-  { icon: <UserRoleIcon fontSize='small' />, path: 'roles', label: 'Roles & Permissions' },
-  { icon: <GroupAddOutlinedIcon fontSize='small' />, path: 'invites', label: 'Invites' },
+  { icon: <UserRoleIcon fontSize='small' />, path: 'roles', label: 'Roles & Permissions', adminOnly: true },
+  { icon: <GroupAddOutlinedIcon fontSize='small' />, path: 'invites', label: 'Invites', adminOnly: true },
   { icon: <FileDownloadOutlinedIcon fontSize='small' />, path: 'import', label: 'Import' },
   { icon: <CloudSyncOutlinedIcon fontSize='small' />, path: 'api', label: 'API' },
   { icon: <CreditCardIcon fontSize='small' />, path: 'subscription', label: 'Billing' },
@@ -27,13 +42,10 @@ export const SPACE_SETTINGS_TABS = [
   { icon: <ProposalIcon fontSize='small' />, path: 'proposals', label: 'Proposals' },
   { icon: <AppsIcon fontSize='small' />, path: 'integrations', label: 'Integrations' },
   { icon: <NotificationsOutlinedIcon fontSize='small' />, path: 'notifications', label: 'Notifications' }
-] as const;
+];
 
-export type SpaceSettingsTab = (typeof SPACE_SETTINGS_TABS)[number];
-export type SpaceSettingsSection = SpaceSettingsTab['path'];
-export const accountSettingsSections = ['account', 'profile', 'charms'] as const;
-
-export type UserSettingsSection = (typeof accountSettingsSections)[number];
+export type UserSettingsSection = 'account' | 'profile' | 'charms';
+export const accountSettingsSections: UserSettingsSection[] = ['account', 'profile', 'charms'];
 
 export const ACCOUNT_TABS = [
   { icon: <ManageAccountsIcon fontSize='small' />, path: 'account', label: 'My Account' },
