+++ components/settings/integrations/IntegrationSettings.tsx
@@ -1,8 +1,8 @@
 import type { Space } from '@charmverse/core/prisma-client';
 
 import { useTrackPageView } from 'charmClient/hooks/track';
+import Legend from 'components/settings/components/Legend';
 import { SpaceIntegrations } from 'components/settings/integrations/components/SpaceIntegrations';
-import Legend from 'components/settings/Legend';
 
 export function IntegrationSettings({ space }: { space: Space }) {
   useTrackPageView({ type: 'settings/integrations' });
