+++ components/settings/proposals/SpaceProposalSettings.tsx
@@ -15,7 +15,7 @@ import { useIsAdmin } from 'hooks/useIsAdmin';
 import { useSnackbar } from 'hooks/useSnackbar';
 import { useSpaceFeatures } from 'hooks/useSpaceFeatures';
 
-import Legend from '../Legend';
+import Legend from '../components/Legend';
 
 import type { WorkflowTemplateFormItem } from './components/ProposalWorkflow';
 import { ProposalWorkflowItem } from './components/ProposalWorkflow';
